requirements.txt
pika==1.3.1
pymongo==4.5.0
fastapi==0.104.1
uvicorn==0.24.0
python-dateutil==2.8.2
datetime==5.3